fix(lint): extend clang-tidy to backends/ root + shared/ impl
Closes the last lint-coverage gap: the backend-agnostic implementation
files (idx.c, log.c, probes.c, shared_utils.c, safetensors.c,
shared/training/*) were clang-tidy'd by nothing. Add them to the tape
clang-tidy pass (LINT_C_CORE_SRCS, -I cJSON for safetensors). clang-tidy
stays implementation-only — examples (backend_byo.c), benches, RTS glue
(refc_shims.c, needs gmp), and test files are out of scope.

Triage of the ~40 surfaced findings:
  - Real fixes: shared_utils.c int*int->wider-type widening (1024L/1024ULL),
    idx.c malloc(0) guard on corrupt/empty dataset dims.
  - NOLINT: 2 idx.c clang-analyzer-unix.Stream FPs (trusted loader; a
    truncated file falls through to the bad-magic/dim check).
  - .clang-tidy disables (inappropriate for this codebase, with rationale):
    android-* (not Android; macOS fopen has no 'e'),
    clang-analyzer-optin.taint.* (taint model assumes untrusted input;
    checkpoints/datasets are trusted — also kills the tainted-index
    ArrayBound), bugprone-multi-level-implicit-pointer-conversion (C
    FFI/JSON char**->void* idiom), rand-family checks (intentional
    seedable non-crypto RNG for dropout), readability-redundant-declaration
    (backend.h FFI surface intentionally overlaps local headers).

+++ packages/backends/shared_utils.c
@@ -21,7 +21,7 @@
 double _wall_ms(void) {
 	struct timeval tv;
 	gettimeofday(&tv, NULL);
-	return tv.tv_sec * 1000.0 + tv.tv_usec / 1000.0;
+	return (double)tv.tv_sec * 1000.0 + (double)tv.tv_usec / 1000.0;
 }
 
 /* --- Seeded per-stream index array (DataStream) ---
@@ -97,7 +97,7 @@ int get_rss_mb(void) {
 	struct rusage usage;
 	getrusage(RUSAGE_SELF, &usage);
 #ifdef __APPLE__
-	return (int)(usage.ru_maxrss / (1024 * 1024));
+	return (int)(usage.ru_maxrss / (1024L * 1024));
 #else
 	return (int)(usage.ru_maxrss / 1024);
 #endif
@@ -114,7 +114,7 @@ int get_current_rss_mb(void) {
 	/* cppcheck-suppress uninitvar */
 	if (task_info(mach_task_self(), MACH_TASK_BASIC_INFO, (task_info_t)&info, &count) ==
 	    KERN_SUCCESS)
-		return (int)(info.resident_size / (1024 * 1024));
+		return (int)(info.resident_size / (1024ULL * 1024));
 #endif
 	return get_rss_mb();
 }
